api/requirements.txt
google-generativeai==0.8.1
requests==2.31.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
